+++ src/components/manage/Widgets/TextWidget.test.jsx
@@ -1,10 +1,24 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
+import configureStore from 'redux-mock-store';
+import { Provider } from 'react-intl-redux';
+
 import TextWidget from './TextWidget';
 
+const mockStore = configureStore();
+
 test('renders a text widget component', () => {
+  const store = mockStore({
+    intl: {
+      locale: 'en',
+      messages: {},
+    },
+  });
+
   const component = renderer.create(
-    <TextWidget id="my-field" title="My field" onChange={() => {}} />,
+    <Provider store={store}>
+      <TextWidget id="my-field" title="My field" onChange={() => {}} />
+    </Provider>,
   );
   const json = component.toJSON();
   expect(json).toMatchSnapshot();